Move logic to MiniBackup class
